requirements.txt
requests
django
pycryptodomex